[ms-cxxabi] Use desugared TSI in the mangler instead of canonical types
At this point, it's clear that the MSVC mangler uses the type-as-written
instead of the canonical type, so this should bring us closer to MSVC.

In particular, this makes the existing array parameter code simpler and
more correct.

Fixes PR16096.

This also fixes a separate issue that Richard spotted in review.
Because separate declarations of the same entity can be spelled and
mangled differently, MSVC always mangles the earliest declaration in an
attempt to avoid link errors.  Clang now does the same.

Differential Revision: http://llvm-reviews.chandlerc.com/D844