[Driver] Start work on an MSVC compatible clang-cl driver
Summary:
We detect this compatibility mode with a flag and based on argv[0]
ending in "cl" for clang-cl.exe or just cl.exe.

Once we've detected this mode, we try to parse arguments as MSVC style
arguments before falling back to parsing gcc style arguments.

Conflicting options can be passed using an -Xclang-driver escape.

High level discussion is on cfe-dev:
http://lists.cs.uiuc.edu/pipermail/cfe-dev/2013-June/030404.html

CC: cfe-commits

Differential Revision: http://llvm-reviews.chandlerc.com/D1048